Several Changes in Strabon Endpoint: Going for a change in the organization. The endpoint shall be composed of the beans covering the functionality of SPARQL protocol and the visual interface. The beans (Describe, Store, Query, and Update) will conform to the SPARQL Protocol specification (e.g., they shall not recognize "format", "SPARQLQuery" parameters, and other Strabon specific). Strabon specific parameters shall only be used in the visual HTML interface. Hence, the Beans will operate on the basis of "Accept" HTML headers.
Now, there is a header for the TELEIOS logo and text description of the Strabon Endpoint. It is included in every .jsp file with a JSP directive. The file is named
"teleios-header.html".

With the introduction of these changes, I hope that nobody could ever mess the code again. :-) Of course, the human potential to messing is enormous.

Still, there is much refactoring to do in order to bring the endpoint at the state conforming to the above description (I have not messed much the functionality of the
visual interface, at least I suppose so. Accept my apologies in the case I have messed it).