Addresses Ticket #20: added support (not completed yet) for requesting the capabilities of Strabon Endpoint
For this a new servlet has been created and can be accessed as /Capabilities. The purpose of this servlet is
to print the capabilities of the Strabon Endpoint implemented by a class that implements the Capabilities
interface. There are two such classes (and might not exist any more): EndpointCapabilities and
AutoDiscoveryCapabilities.

- EndpointCapabilities should be used for endpoints with version >= 3.2.5 (i.e., the next release of Strabon
Endpoint).
- AutoDiscoveryCapabilities should be used for old endpoints with version <= 3.2.4. The purpose of this
implementation is to attempt to find out the capabilities of old Strabon Endpoints based on java reflection
and response messages or HTTP codes of service requests. The result might not be always accurate.
Discovery of how a service can be accessed might be done using HttpClient and catching error codes.

A new bean (like StrabonBeanWrapper) has also been created, named CapabilitiesDelegateBean, which is
instantiated with one of the Capabilities implementations, i.e., either EndpointCapabilities or
AutoDiscoveryCapabilities class. Versions <= 3.2.4 should be instantiated using the AutoDiscoveryCapabilities
implementation.

Implementation concerning this feature has been transferred in package
eu.earthobservatory.org.StrabonEndpoint.capabilities. This is to facilitate exporting this package as a jar
and then placed inside the WEB-INF/lib directory of an old endpoint in order to determine its capabilities.

The procedure is the following:
1) Extract package eu.earthobservatory.org.StrabonEndpoint.capabilities as a jar and place it inside
   directory WEB-INF/lib of an old endpoint,
2) modify beans.xml so as tomcat instantiates the bean CapabilitiesDelegateBean with auto discover
   functionality by adding the following lines:

	<bean id="capsBean" class="eu.earthobservatory.org.StrabonEndpoint.capabilities.CapabilitiesDelegateBean" destroy-method="destroy">
		<!-- specify which implementation of the Capabilities interface shall be used -->
		<!-- true : AutoDiscoveryCapabilities -->
		<!-- false: EndpointCapabilities -->
		<constructor-arg type="boolean">
            <value>true</value>
        </constructor-arg>
	</bean>

3) modify web.xml to define a Capabilities servlet adding the following lines:

	<servlet>
		<display-name>Capabilities</display-name>
		<servlet-name>Capabilities</servlet-name>
		<servlet-class>eu.earthobservatory.org.StrabonEndpoint.capabilities.CapabilitiesBean</servlet-class>
		<load-on-startup>1</load-on-startup>
	</servlet>
	<servlet-mapping>
		<servlet-name>Capabilities</servlet-name>
		<url-pattern>/Capabilities</url-pattern>
	</servlet-mapping>

For future development of services in Strabon, notice that when a new service (XXXBean) is created, a corresponding class
XXXCapabilities should also be created in package eu.earthobservatory.org.StrabonEndpoint.capabilities.